Moved all time slots from Store reducer to Delivery reducer
