+++ src/redux/Delivery/actions.js
@@ -1,6 +1,8 @@
+import _ from 'lodash';
 import { createAction } from "@reduxjs/toolkit";
 
 import { setLoading } from '../App/actions';
+import { selectTimeSlots } from "./selectors";
 
 
 export const SET_RETURN_SCREEN = '@delivery/SET_RETURN_SCREEN';
@@ -11,6 +13,7 @@ export const LOAD_ADDRESSES_SUCCESS = '@delivery/LOAD_ADDRESSES_SUCCESS';
 
 export const LOAD_TIME_SLOTS_SUCCESS = '@delivery/LOAD_TIME_SLOTS_SUCCESS';
 export const LOAD_TIME_SLOT_SUCCESS = '@delivery/LOAD_TIME_SLOT_SUCCESS';
+export const LOAD_TIME_SLOT_CHOICES_SUCCESS = '@delivery/LOAD_TIME_SLOT_CHOICES_SUCCESS';
 
 export const SET_REFRESHING = '@delivery/SET_REFRESHING';
 
@@ -34,6 +37,7 @@ const getPriceError = createAction(GET_PRICE_ERROR);
 
 const loadTimeSlotsSuccess = createAction(LOAD_TIME_SLOTS_SUCCESS);
 const loadTimeSlotSuccess = createAction(LOAD_TIME_SLOT_SUCCESS);
+export const loadTimeSlotChoicesSuccess = createAction(LOAD_TIME_SLOT_CHOICES_SUCCESS);
 
 
 const loadAddressesSuccess = createAction(
@@ -84,6 +88,79 @@ export function loadAddresses(store) {
       .catch(e => {
         dispatch(setLoading(false));
         dispatch(setRefreshing(false));
+      });
+  };
+}
+
+
+
+export function loadTimeSlots(store) {
+  return (dispatch, getState) => {
+    const { app } = getState();
+    const { httpClient } = app;
+
+    dispatch(setLoading(true));
+
+    return httpClient
+      .get(`${store['@id']}/time_slots`)
+      .then(res => {
+        dispatch(loadTimeSlotsSuccess(res['hydra:member']));
+        dispatch(setLoading(false));
+        // dispatch(loadTimeSlotChoices(res['hydra:member'][0]));
+      })
+      .catch(e => {
+        dispatch(setLoading(false));
+      });
+  };
+}
+
+
+export function loadTimeSlotChoices(timeSlot) {
+  return (dispatch, getState) => {
+    const { app } = getState();
+    const { httpClient } = app;
+
+    dispatch(setLoading(true));
+
+    return httpClient
+      .get(`${timeSlot['@id']}/choices`)
+      .then(res => {
+        dispatch(loadTimeSlotChoicesSuccess(res.choices));
+        dispatch(setLoading(false));
+      })
+      .catch(e => {
+        dispatch(setLoading(false));
+      });
+  };
+}
+
+export function loadTimeSlot(store) {
+  return (dispatch, getState) => {
+    if (!store.timeSlot) {
+      return;
+    }
+
+    const { app } = getState();
+    const { httpClient } = app;
+
+    const timeSlot = _.find(
+      selectTimeSlots(getState()),
+      ts => ts['@id'] === store.timeSlot,
+    );
+    if (timeSlot) {
+      return;
+    }
+
+    dispatch(setLoading(true));
+
+    return httpClient
+      .get(store.timeSlot)
+      .then(res => {
+        dispatch(setLoading(false));
+        dispatch(loadTimeSlotSuccess(res));
+      })
+      .catch(e => {
+        dispatch(setLoading(false));
       });
   };
 }
