+++ src/redux/Delivery/reducers.js
@@ -5,6 +5,7 @@ import {
   GET_PRICE_ERROR,
   GET_PRICE_SUCCESS,
   LOAD_ADDRESSES_SUCCESS,
+  LOAD_TIME_SLOTS_SUCCESS,
   SET_RETURN_SCREEN,
   SET_STORE
 } from "./actions"
