+++ src/redux/Store/actions.js
@@ -3,7 +3,6 @@ import { createAction } from 'redux-actions';
 
 import { setLoading } from '../App/actions';
 import { selectStore } from './selectors';
-import { selectTimeSlots } from '../Delivery/selectors';
 
 export const LOAD_DELIVERIES_SUCCESS = '@store/LOAD_DELIVERIES_SUCCESS';
 export const CREATE_DELIVERY_SUCCESS = '@store/CREATE_DELIVERY_SUCCESS';
@@ -11,8 +10,6 @@ export const LOAD_TASKS_SUCCESS = '@store/LOAD_TASKS_SUCCESS';
 export const SET_LOADING_MORE = '@store/SET_LOADING_MORE';
 export const SET_REFRESHING = '@store/SET_REFRESHING';
 export const INIT_SUCCESS = '@store/INIT_SUCCESS';
-export const LOAD_TIME_SLOT_CHOICES_SUCCESS =
-  '@store/LOAD_TIME_SLOT_CHOICES_SUCCESS';
 export const LOAD_PACKAGES_SUCCESS = '@store/LOAD_PACKAGES_SUCCESS';
 
 export const loadPackagesSuccess = createAction(LOAD_PACKAGES_SUCCESS);
@@ -35,46 +32,6 @@ export function loadPackages(store) {
   };
 }
 
-export function loadTimeSlots(store) {
-  return (dispatch, getState) => {
-    const { app } = getState();
-    const { httpClient } = app;
-
-    dispatch(setLoading(true));
-
-    return httpClient
-      .get(`${store['@id']}/time_slots`)
-      .then(res => {
-        dispatch(loadTimeSlotsSuccess(res['hydra:member']));
-        dispatch(setLoading(false));
-        // dispatch(loadTimeSlotChoices(res['hydra:member'][0]));
-      })
-      .catch(e => {
-        dispatch(setLoading(false));
-      });
-  };
-}
-export const loadTimeSlotChoicesSuccess = createAction(
-  LOAD_TIME_SLOT_CHOICES_SUCCESS,
-);
-export function loadTimeSlotChoices(timeSlot) {
-  return (dispatch, getState) => {
-    const { app } = getState();
-    const { httpClient } = app;
-
-    dispatch(setLoading(true));
-
-    return httpClient
-      .get(`${timeSlot['@id']}/choices`)
-      .then(res => {
-        dispatch(loadTimeSlotChoicesSuccess(res.choices));
-        dispatch(setLoading(false));
-      })
-      .catch(e => {
-        dispatch(setLoading(false));
-      });
-  };
-}
 export const createDeliverySuccess = createAction(CREATE_DELIVERY_SUCCESS);
 export const loadDeliveriesSuccess = createAction(
   LOAD_DELIVERIES_SUCCESS,
@@ -215,37 +172,6 @@ export function loadTasks(delivery) {
       })
       .catch(e => {
         console.log(e);
-        dispatch(setLoading(false));
-      });
-  };
-}
-
-export function loadTimeSlot(store) {
-  return (dispatch, getState) => {
-    if (!store.timeSlot) {
-      return;
-    }
-
-    const { app } = getState();
-    const { httpClient } = app;
-
-    const timeSlot = _.find(
-      selectTimeSlots(getState()),
-      ts => ts['@id'] === store.timeSlot,
-    );
-    if (timeSlot) {
-      return;
-    }
-
-    dispatch(setLoading(true));
-
-    return httpClient
-      .get(store.timeSlot)
-      .then(res => {
-        dispatch(setLoading(false));
-        dispatch(loadTimeSlotSuccess(res));
-      })
-      .catch(e => {
         dispatch(setLoading(false));
       });
   };
