+++ src/redux/Store/reducers.js
@@ -4,7 +4,6 @@ import {
   LOAD_DELIVERIES_SUCCESS,
   LOAD_PACKAGES_SUCCESS,
   LOAD_TASKS_SUCCESS,
-  LOAD_TIME_SLOTS_SUCCESS,
   LOAD_TIME_SLOT_CHOICES_SUCCESS,
   SET_LOADING_MORE,
   SET_REFRESHING,
